WIP Using internal storage - Refactor Workspace usage
