+++ app/src/main/java/org/sil/storyproducer/activities/BaseActivity.kt
@@ -32,10 +32,6 @@ import java.io.File
 
 open class BaseActivity : AppCompatActivity(), BaseActivityView {
 
-    val NUM_BYTES_NEEDED_FOR_DEMO_STORY = 1024 * 1024 * 10L;
-    val SP_TEMPLATES_FOLDER_NAME = "SP Templates"
-    private val autoInternalWorkspace = true
-
     lateinit var controller: SelectTemplatesFolderController
 
     private var readingTemplatesDialog: AlertDialog? = null
@@ -72,37 +68,9 @@ open class BaseActivity : AppCompatActivity(), BaseActivityView {
         if (Workspace.workdocfile.isDirectory) {
             controller.updateStories()  // We already have a workspace folder
             showWelcome = false
-        } else if (autoInternalWorkspace) {
-            // Check App specific internal or external storage volumes for free space for a new workspace demo folder
-            // https://developer.android.com/training/data-storage
-            // https://developer.android.com/training/data-storage/app-specific
-            var appSpecificDocsDir : File? = null
-            var appSpecificTemplateDir : File? = null
-            val freeIntMem = getFreeInternalMemorySize()
-            val freeExtMem = getFreeExternalMemorySize()
-            // Check external memory first then internal memory for enough free space for demo
-            if (freeExtMem >= NUM_BYTES_NEEDED_FOR_DEMO_STORY) {
-                appSpecificDocsDir = App.appContext.getExternalFilesDir(Environment.DIRECTORY_DOCUMENTS)
-                if (appSpecificDocsDir != null)
-                    appSpecificTemplateDir = File(appSpecificDocsDir, SP_TEMPLATES_FOLDER_NAME)
-            } else if (freeIntMem >= NUM_BYTES_NEEDED_FOR_DEMO_STORY) {
-                appSpecificDocsDir = App.appContext.filesDir
-                appSpecificTemplateDir = File(appSpecificDocsDir, SP_TEMPLATES_FOLDER_NAME)
-            }
-            if (appSpecificTemplateDir != null) {
-                // we have space on this volume for a demo story
-                appSpecificTemplateDir.mkdirs()
-                if (appSpecificTemplateDir.isDirectory) {
-                    // if we managed to create an internal workspace folder
-                    if (Workspace.setupWorkspacePath(App.appContext, Uri.fromFile(appSpecificTemplateDir))) {
-                        Workspace.isInitialized = true  // internal workspace successful
-                        // Add the demo for new users
-                        Workspace.addDemoToWorkspace(App.appContext)
-                        controller.updateStories()  // always refresh list of stories
-                        showWelcome = false
-                    }
-                }
-            }
+        } else if (Workspace.createAppSpecificWorkspace()) {
+            controller.updateStories()  // Update after a new app-specific workspace was created
+            showWelcome = false
         }
         if (showWelcome) {
             showWelcomeDialog()
